refactor(payment-destination): share a send-destination guard and relocate username resolution
