+++ app/self-custodial/payment-details/wrap-destination.ts
@@ -6,12 +6,11 @@ import { WalletCurrency } from "@app/graphql/generated"
 import { FeeTierOption } from "@app/screens/send-bitcoin-screen/hooks/fee-tiers.types"
 import {
   type CreatePaymentDetailParams,
-  DestinationDirection,
+  isSendDestination,
   type ParseDestinationResult,
 } from "@app/screens/send-bitcoin-screen/payment-destination/index.types"
 import { toBtcMoneyAmount } from "@app/types/amounts"
 import { PaymentType as SelfCustodialPaymentType } from "@app/types/transaction"
-import { getLightningAddress } from "@app/utils/pay-links"
 
 import { createSelfCustodialLightningPaymentDetails } from "./lightning"
 import { createSelfCustodialLnurlPaymentDetails } from "./lnurl"
@@ -116,8 +115,7 @@ export const wrapDestination = (
   sdk: BreezSdkInterface,
   options: WrapOptions = {},
 ): ParseDestinationResult => {
-  if (!result.valid) return result
-  if (result.destinationDirection !== DestinationDirection.Send) return result
+  if (!isSendDestination(result)) return result
 
   const original = result.validDestination
 
@@ -160,29 +158,4 @@ export const wrapDestination = (
       }
     },
   }
-}
-
-const intraledgerHandle = (result: ParseDestinationResult): string | undefined => {
-  if (!result.valid || result.destinationDirection !== DestinationDirection.Send) {
-    return undefined
-  }
-  const { validDestination } = result
-  if (
-    validDestination.paymentType === PaymentType.Intraledger ||
-    validDestination.paymentType === PaymentType.IntraledgerWithFlag
-  ) {
-    return validDestination.handle
-  }
-  return undefined
-}
-
-export const resolveUsername = async (
-  result: ParseDestinationResult,
-  lnAddressHostname: string,
-  resolveLnAddress: (rawInput: string) => Promise<ParseDestinationResult>,
-): Promise<ParseDestinationResult> => {
-  const handle = intraledgerHandle(result)
-  return handle
-    ? resolveLnAddress(getLightningAddress(lnAddressHostname, handle))
-    : result
 }
